rename custom to extended, fits the actual abilities of the API better also move the packet ID out of the interface, no reason to be API
