+++ library/entity/entity_networking/src/main/java/org/quiltmc/qsl/entity/networking/impl/QuiltEntityNetworkingInitializer.java
@@ -32,6 +32,8 @@ import org.quiltmc.qsl.registry.api.sync.RegistrySynchronization;
 
 @ApiStatus.Internal
 public class QuiltEntityNetworkingInitializer implements ModInitializer {
+	public static final Identifier EXTENDED_SPAWN_PACKET = new Identifier("quilt", "extended_entity_spawn_packet");
+
 	public static final SimpleRegistry<TrackedDataHandler<?>> TRACKED_DATA_HANDLER_REGISTRY = new SimpleRegistry<>(
 			RegistryKey.ofRegistry(new Identifier("quilt", "tracked_data_handlers")), Lifecycle.stable(), false
 	);
